requirements.txt
fastapi
uvicorn
httpx
aiohttp
python-dotenv
